refactor(cqrs): move 4 bus implementation files into buses/ subfolder
Move command-bus.ts, event-bus.ts, query-bus.ts, saga-bus.ts into
src/buses/ to separate bus implementations from core module files,
per docs/reference/package-folder-structure.md convention.

Update all internal import paths within the package. Public API
re-export signatures in index.ts remain identical.

Closes #853
